fix nonexistant variable calls and add null filter to `withChildren`
I removed calls for `String.isString()` and replaced them with the appropriate tests. Additionally, there was a leftover namespace call for the Elmen class; this has been fixed to be just “Elmen.” I also added a null filter to `withChildren`, so any null or undefined arguments passed into it are ignored. This makes it easier to have a variable number of children.