I removed support for 3-tuple token lists in withCSS.
It never really made sense to have the option for switching based on whether the list length is divisible by two or three, because it can be both. Making it just two seems like the right choice, because “!important” is rarely used anyway, and leaving a spot for it every time is a huge waste of space, especially considering that you almost never need that for inline styles.